first commit in kosmw-admin using tsx and jsx and js files
Layout is the main layout for all the main page display
NavBar is the side bar that is fixed
Productform is handling all the front end input info in order to create or edit a product.
UploadComponent is handling the upload of the info including the link of the images.
db is the connection to the database
mongoose using mongoose to make the connection
[next-auth]: GoogleProvider so far
api/bulk-delete: bulk delete of the products in product page
api/products: the api connection to the database put/post/delete/postall
api/upload: api connection to aws s3 to upload images and get the links
pages/products:  products add new and select and delete
settings and orders -- To- Be-Continued
products/delete. handling the delete of the of one product but not in use
products/edit. handling the front end of editing the page
products/new handling the adding a new product
sign-in -- not-in-use