Images
changed the logic of which images are displayed in productGallery and how they are uploaded to db. Only uploaded images will be uploaded to the product so this issue is fixed here. Added the logic that if the product hasnt changed and we press upload we just go back to products. I added useRef to check the current state with the updated one.